refactor(next-query): keep parsing policy in SparqlParser
